requirements.txt
hydra-core==1.3.1
matplotlib==3.6.3
omegaconf==2.3.0
Pillow==9.4.0
pytest==7.2.1
pytorch_lightning==1.9.2
scikit_learn==1.2.1
setuptools==67.1.0
torchvision==0.14.1
